+++ spec/integration/ticket_priority_spec.rb
@@ -99,7 +99,7 @@ RSpec.describe ZammadAPI, 'ticket priority object basics' do
   it 'pagination with all' do
     ticket_priorities = client.ticket_priority.all
 
-    expect(ticket_priorities[0].class).to eq(ZammadAPI::Resources::TicketPriority)
+    expect(ticket_priorities.first.class).to eq(ZammadAPI::Resources::TicketPriority)
 
     count = 0
     ticket_priorities.each do |local_ticket_priority|
@@ -110,17 +110,17 @@ RSpec.describe ZammadAPI, 'ticket priority object basics' do
 
     count = 0
     ticket_priorities = client.ticket_priority.all
-    ticket_priorities.page(1, per_page: 2).each do |local_ticket_priority|
+    ticket_priorities.page(1).per(2).each do |local_ticket_priority|
       expect(local_ticket_priority.class).to eq(ZammadAPI::Resources::TicketPriority)
       count += 1
     end
     expect(count).to eq(2)
-    ticket_priorities.page(2, per_page: 2).each do |local_ticket_priority|
+    ticket_priorities.page(2).per(2).each do |local_ticket_priority|
       expect(local_ticket_priority.class).to eq(ZammadAPI::Resources::TicketPriority)
       count += 1
     end
     expect(count).to eq(4)
-    ticket_priorities.page(3, per_page: 2).each do |local_ticket_priority|
+    ticket_priorities.page(3).per(2).each do |local_ticket_priority|
       expect(local_ticket_priority.class).to eq(ZammadAPI::Resources::TicketPriority)
       count += 1
     end